+++ src/main/data/hullmods/BaseVenting.java
@@ -5,6 +5,7 @@ import com.fs.starfarer.api.Global;
 import com.fs.starfarer.api.combat.CombatEngineAPI;
 import com.fs.starfarer.api.combat.ShipAPI;
 import com.fs.starfarer.api.combat.ShipCommand;
+import data.combatlog.Util;
 
 import java.util.Map;
